+++ game_patch/input/mouse.h
@@ -2,3 +2,4 @@
 
 void mouse_apply_patch();
 void consume_raw_mouse_deltas(float& out_pitch, float& out_yaw, bool apply_scope_sens);
+void mouse_handle_xbutton_wm(int rf_btn, bool down);